Added cleaned version of Miranda data file
This is a manually cleaned version of the miranda data done externally to this package.  While this defeats the purpose of this package somewhat, which was to computationally document the cleaning process starting from the source data, in this case that ideal was impossible due to a great many inconsistencies including dialog and commentary in non-standard cells direction actions to be taken on other cells, and color-coded groupings of cells meant to indicate quality, actions (change or replace rows/cells), or "copy down" type cases where the above data was meant to apply to lower rows as well.